Use updated API for setting display mode
The new version of KaTeX allows passing an options hash to `render` for
setting the `displayMode` property to a boolean value. Perviously we
needed to store the original expression and optionally wrap it in
delimiters to control the display mode. Now we can simply use the
boolean `display` attribute directly.